Move lib/slack/middleware to lib/middleware
