+++ lib/commands/index.js
@@ -1,6 +1,6 @@
 // Slash Commands - https://api.slack.com/slash-commands
 
-const middleware = require('../slack/middleware');
+const middleware = require('../middleware');
 const subscribe = require('./subscribe');
 const signin = require('./signin');
 const listSubscriptions = require('./list-subscriptions');
