+++ lib/middleware/authenticate.js
