+++ lib/middleware/get-installation.js
@@ -1,5 +1,5 @@
-const { InstallGitHubApp, NotFound } = require('../../messages/flow');
-const SignedParams = require('../../signed-params');
+const { InstallGitHubApp, NotFound } = require('../messages/flow');
+const SignedParams = require('../signed-params');
 
 /**
  * Get the installation for the given account name
