+++ lib/middleware/get-resource.js
@@ -1,6 +1,6 @@
-const githubUrl = require('../../github-url');
+const githubUrl = require('../github-url');
 
-const { InvalidUrl } = require('../../messages/flow');
+const { InvalidUrl } = require('../messages/flow');
 
 /**
  * Looks up the GitHub resource that is being referenced in a command.
