+++ lib/middleware/pending-command.js
