+++ lib/middleware/route-command.js
@@ -1,4 +1,4 @@
-const Command = require('../command');
+const Command = require('../slack/command');
 
 /**
  * Parses subcommands and routes the request based on the subcommand
