+++ lib/middleware/sentry-context.js
