+++ lib/slack/index.js
@@ -5,7 +5,7 @@ const commands = require('../commands');
 const oauth = require('./oauth');
 const importer = require('./importer');
 const uninstall = require('./uninstall');
-const middleware = require('./middleware');
+const middleware = require('./../middleware');
 
 module.exports = (robot) => {
   const router = robot.route('/slack');
